refactor(entityNotifications): refactor export
