+++ app/scripts/modules/core/src/entityTag/entityTags.module.ts
@@ -2,7 +2,7 @@ import { module } from 'angular';
 
 import { ADD_ENTITY_TAG_LINKS_COMPONENT } from './addEntityTagLinks.component';
 import { DATA_SOURCE_NOTIFICATIONS } from './notifications/DataSourceNotifications';
-import { ENTITY_NOTIFICATIONS } from './notifications/EntityNotifications';
+import { ENTITY_NOTIFICATIONS } from './notifications/entityNotifications.component';
 import { ENTITY_SOURCE_COMPONENT } from './entitySource.component';
 import { ENTITY_TAG_WRITER } from './entityTags.write.service';
