+++ app/scripts/modules/core/src/entityTag/index.ts
@@ -3,4 +3,3 @@ export * from './clusterTargetBuilder.service';
 export * from './entityTags.read.service';
 export * from './entityTags.write.service';
 export * from './notifications';
-
