+++ app/scripts/modules/core/src/entityTag/notifications/EntityNotifications.tsx
@@ -1,6 +1,3 @@
-import { module } from 'angular';
-
-import { react2angular } from 'react2angular';
 import * as React from 'react';
 import autoBindMethods from 'class-autobind-decorator';
 
@@ -98,37 +95,3 @@ export class EntityNotifications extends React.Component<IEntityNotificationsPro
     );
   }
 }
-
-
-export const ENTITY_NOTIFICATIONS = 'spinnaker.core.entityTag.alerts.entitynotifications';
-const ngmodule = module(ENTITY_NOTIFICATIONS, []);
-
-ngmodule.component('entityNotificationsWrapper', react2angular(EntityNotifications, [
-  'entity', 'application', 'placement', 'hOffsetPercent', 'className', 'pageLocation', 'entityType', 'onUpdate'
-]));
-
-
-ngmodule.component('entityNotifications', {
-  template: `
-    <entity-notifications-wrapper
-      entity="$ctrl.entity"
-      application="$ctrl.application"
-      placement="$ctrl.placement"
-      h-offset-percent="$ctrl.hOffsetPercent"
-      class-name="$ctrl.className"
-      entity-type="$ctrl.entityType"
-      page-location="$ctrl.pageLocation"
-      on-update="$ctrl.onUpdate"
-    ></entity-notifications-wrapper>
-  `,
-  bindings: {
-    entity: '<',
-    application: '<',
-    placement: '@',
-    hOffsetPercent: '@',
-    className: '@',
-    entityType: '@',
-    pageLocation: '@',
-    onUpdate: '&',
-  }
-});
