+++ app/scripts/modules/core/src/entityTag/notifications/index.ts
@@ -1,6 +1,5 @@
 export * from './CategorizedNotifications';
 export * from './EntityName';
-export * from './EntityNotifications';
 export * from './GroupedNotificationList';
 export * from './NotificationCategory';
 export * from './NotificationList';
